[bug] 보안문자 이동 페이지 수정 #64
